requirements.txt
astroid==2.13.2
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dill==0.3.6
flake8==6.0.0
fonttools==4.38.0
idna==3.4
isort==5.11.4
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.6.2
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.15.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
